Move Chronogram code into its own directory
